Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 6ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250608/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -11,7 +11,7 @@ int main() {
     cin >> m; // [1,m] 정답
     cin >> a >> b; // a,...,b시 count 최대 최소 
 
-    int min_value = INT_MAX;
+    int min_value = LLONG_MAX;
     int max_value = 0;
 
     for (int target = a; target <= b; target++) {
